refactor(imports): move _types files to index.ts
