+++ apps/app/src/lib/query/_queries/download.ts
@@ -1,6 +1,6 @@
 import * as services from '$lib/services';
 import type { Recording } from '$lib/services/db';
-import type { DownloadServiceError } from '$lib/services/download/_types';
+import type { DownloadServiceError } from '$lib/services/download';
 import { Err, type Result } from '@epicenterhq/result';
 import type { WhisperingError } from '$lib/result';
 import { defineMutation } from '../_utils';
