+++ apps/app/src/lib/services/clipboard/desktop.ts
@@ -2,7 +2,7 @@ import { Err, Ok, tryAsync, type Result } from '@epicenterhq/result';
 import { invoke } from '@tauri-apps/api/core';
 import { writeText } from '@tauri-apps/plugin-clipboard-manager';
 import { type } from '@tauri-apps/plugin-os';
-import type { ClipboardService, ClipboardServiceError } from './_types';
+import type { ClipboardService, ClipboardServiceError } from '.';
 import type { WhisperingWarning } from '$lib/result';
 
 export function createClipboardServiceDesktop(): ClipboardService {
