+++ apps/app/src/lib/services/os/desktop.ts
@@ -1,5 +1,5 @@
-import type { OsService } from './types';
 import * as os from '@tauri-apps/plugin-os';
+import type { OsService } from '.';
 
 export function createOsServiceDesktop(): OsService {
 	return {
