+++ apps/app/src/lib/services/sound/web.ts
@@ -1,6 +1,6 @@
 import { Err, Ok } from '@epicenterhq/result';
-import { extension } from '@repo/extension';
-import type { PlaySoundService } from './_types';
+// import { extension } from '@repo/extension';
+import type { PlaySoundService } from '.';
 import { audioElements } from './_audioElements';
 
 export function createPlaySoundServiceWeb(): PlaySoundService {
@@ -10,17 +10,17 @@ export function createPlaySoundServiceWeb(): PlaySoundService {
 				await audioElements[soundName].play();
 				return Ok(undefined);
 			}
-			const { error: playSoundError } = await extension.playSound({
-				sound: soundName,
-			});
-			if (playSoundError) {
-				return Err({
-					name: 'PlaySoundServiceError',
-					message: `We encountered an issue while playing the ${soundName} sound`,
-					context: { soundName },
-					cause: playSoundError,
-				});
-			}
+			// const { error: playSoundError } = await extension.playSound({
+			// 	sound: soundName,
+			// });
+			// if (playSoundError) {
+			// 	return Err({
+			// 		name: 'PlaySoundServiceError',
+			// 		message: `We encountered an issue while playing the ${soundName} sound`,
+			// 		context: { soundName },
+			// 		cause: playSoundError,
+			// 	});
+			// }
 			return Ok(undefined);
 		},
 	};
